Refactor PageTitle
