+++ src/components/publiq-ui/Page.stories.mdx
@@ -1,6 +1,5 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
 import { Page } from './Page.js';
-import { PageTitle } from './PageTitle.js';
 
 <Meta title="Components/Page" />
 
@@ -22,7 +21,7 @@ import { PageTitle } from './PageTitle.js';
   <Story name="WithPageTitle">
     {(args) => (
       <Page {...args}>
-        <PageTitle>Page Title</PageTitle>
+        <Page.Title>Page Title</Page.Title>
         <div>Page content</div>
       </Page>
     )}
