--- src/components/publiq-ui/PageTitle.stories.mdx
@@ -1,12 +1,0 @@
-import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
-import { PageTitle } from './PageTitle.js';
-
-<Meta title="Components/PageTitle" />
-
-<Canvas>
-  <Story name="Default" args={{ children: 'Page title' }}>
-    {(args) => <PageTitle {...args} />}
-  </Story>
-</Canvas>
-
-<ArgsTable story="Default" />
