+++ src/pages/dashboard.js
@@ -1,10 +1,9 @@
 import { Page } from '../components/publiq-ui/Page';
-import { PageTitle } from '../components/publiq-ui/PageTitle';
 
 const Dashboard = () => {
   return (
     <Page>
-      <PageTitle>Dashboard</PageTitle>
+      <Page.Title>Dashboard</Page.Title>
     </Page>
   );
 };
